Download OpenStreetMap and Aqueduct data
Dataset is now specified in the config by a URL which is retrieved automatically. Hazard datasets are specified as a named list of .txt files that contain the hazard rasters.

Multiple hazard and infrastructure datasets can be specified in the config. A final output .geoparquet file will be created for each infrastructure-hazard combination.

Output is collected into subdirectories, named for the relevant hazard/infrastructure datasets combination.