Parallelise electricity_grid_disruption with multiprocessing
The previous commit removed parallelism through jobs (single storm per
job). This moves parallelism into each job (running a whole sample of
storms per job) and uses a multiprocessing pool to process the storms.

Note that we use the fork method for spawning subprocesses. This is
available on POSIX systems. It can have problems with deadlocking
when the parent process is threaded. However, usefully, it employs
copy-on-write, meaning if the child processes do not modify the objects
in the parent's memory, the objects can be shared. Potentially meaning
lower memory consumption.