Taper advective wind field as function of radius to eye
The estimated wind field for each time step is a vector sum of two
components, the advective field due to the translational motion of the
storm and the rotational field. The rotational field intensity is a
function of radius from storm eye. Previously, the advective field
had no spatial dependence. This meant the entire spatial domain had some
sort of background wind field (typically(!) a few m/s). In the event of
a fast moving storm (say a decayed cyclone transitting the NA, this
speed may be quite significant. For large domains like the USA, this
could mean not insignificant background wind speeds many thousands of
miles from the storm. I have decided that preventing this is worth the
necessary assumptions and extra complexity.

We now take the advective wind vector and multiply it by some reduction
factor. This factor is the distance to storm eye, normalised by the
radius to maximum winds, then tapered to 0 using a hyperbolic tangent
function.