Reorganise scripts and rules
- workflow rules grouped by high-level concept and workflow stage
- scripts sit next to the rules that use them

Aim to reduce cognitive overhead and sense that different sub-workflows
were muddled together in the previous grouping by general stage.

Aim for a clear set of workflow/<stage> directories, where we should
be able to draw a simple high-level DAG for the workflow between
stages, and sets of rules within each stage are roughly linear.